Correção ao excluir uma conta bancária
Em caso de erro, não estava exibindo nenhum mensagem pro usuário.

+++ lib/features/account/presentation/ui/account_details_page.dart
@@ -204,6 +204,8 @@ class _AccountDetailsPageState extends State<AccountDetailsPage> with ThemeConte
 
       await notifier.delete();
 
+      if (notifier.value is ErrorAccountState) throw Exception((notifier.value as ErrorAccountState).message);
+
       if (!mounted) return;
       context.pop(FormResultNavigation<AccountEntity>.delete());
     } catch (e) {
